added moneyables module, all tests passing
