+++ Person.rb
@@ -1,7 +1,9 @@
 class Person
+	attr_reader :age, :phase
+	attr_accessor :name
 
- def initialize(options={})
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-  end
+	def initialize(options={})
+		@age = options.fetch(:age, 0)
+		@name = options.fetch(:name, "")
+	end
 end
