+++ funables.rb
@@ -1,5 +1,4 @@
 module Funables
-
 	def offer_high_five
 		"High five!"
 	end
