+++ moneyables.rb
@@ -1,0 +1,11 @@
+module Moneyables
+
+def salary=(new_salary)
+    puts "This better be good!"
+    @salary = new_salary
+  end
+
+  def receive_raise(raise)
+    @salary += raise
+  end
+end
