+++ teacher.rb
@@ -1,11 +1,11 @@
 require_relative 'funables'
 require_relative 'person'
 require_relative 'teachables'
-
+require_relative 'moneyables'
 
 class Teacher < Person
-  attr_reader :age, :salary, :phase, :performance_rating, :target_raise
-  attr_accessor :name
+  attr_reader  :salary, :performance_rating, :target_raise
+
 
   def initialize(options={})
     super
@@ -13,12 +13,8 @@ class Teacher < Person
     @target_raise = 1000
   end
 
-  include Funables, Teachables
+  include Funables, Teachables, Moneyables
 
-  # def set_phase(num)
-  #   @phase = num
-  #   "Cool, I've always wanted to teach phase #{num}!"
-  # end
 
   def teach_stuff
     response = ""
@@ -26,15 +22,6 @@ class Teacher < Person
     response += "*drops flat-out insane knowledge bomb* "
     response += "... You're welcome. *saunters away*"
     response
-  end
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
   end
 
   def set_performance_rating(rating)
